#  [bounty] Fix TestOps.test_avg_pool3d_failure by fixing the issue in the linearizer, not with workarounds