@geohot 2025-04-26
Hmm #10056 also fixes it, no linearizer changes. Do you know why?

In general, I really prefer solutions that look like bugfixes and not addition of new codepaths and checks.