@0guanhua0 2025-04-29
core: optimization level

```
acc = 0
for i in range(8):
    acc = 0
    for j in range(8):
        acc = (acc + 1)
print("acc:", acc)

for i in range(8):
    acc1 = 0
    for j in range(8):
        acc1 = (acc1 + 1)
print("acc1:", acc1)
```

```
#include <stdio.h>
int main (void)
{
    int acc = 0;
    for (int i = 0; i < 8; ++i)
    {
        acc = 0;
        for (int j = 0; j < 8; ++j)
        {
            acc = acc + 1;
        }
    }
    printf("acc: %d\n", acc);

    for (int i = 0; i < 8; ++i)
    {
        int acc1 = 0;
        for (int j = 0; j < 8; ++j)
        {
            acc1 = acc1 + 1;
        }
    }
    printf("acc1: %d\n", acc1);

    return 0;
}
```

data: yes, same val
instruction: no, diff scope