@0guanhua0 2025-04-29
before
```
REDUCE_AXIS -> DEFINE_ACC + ASSIGN
```

https://github.com/tinygrad/tinygrad/pull/9527 add Ops.REDUCE
after
```
REDUCE_AXIS -> REDUCE -> DEFINE_ACC + ASSIGN
```

optimize at data level(DEFINE_ACC + ASSIGN) add scope err, so linearizer fail
root fix is to check upstream level(REDUCE)