@0guanhua0 2025-04-29
https://github.com/tinygrad/tinygrad/pull/10052/files#diff-fdbfd74d8cf695c22f1cd280a82554192b82af87324067afc80527c6eacde737R409 hack

https://github.com/tinygrad/tinygrad/pull/10056/files#diff-c8dafc2ec8b46a2e6884917fc9645380e0839da44700ef5c52eae862aaeb2a02R408-R409 deep dive