@0guanhua0 2025-04-29
good example to do the right thing at right time/level/abstraction